todobackend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
pydantic==2.12.5
pydantic-settings==2.6.1
pydantic[email]==2.12.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
apscheduler==3.10.4
fastapi-mail==1.6.1